Allow shallow clones again.
shallow=false was needed previously because the version is set based on
the annotated git tags. However, during rnp's initial CMake switchover,
I added support for 0.0.0 versions when no tags are present.

For releases, we may want shallow=false or some other mechanism, but
currently this is a HEAD-only formula.